python-llm/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.7.2
pydantic==2.9.2

